backend/requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
pytest
httpx
einops
google-genai
python-dotenv
